+++ CHANGELOG.rst
@@ -10,6 +10,29 @@ Unreleased changes are documented in files in the `changelog.d`_ directory.
 
 ..  scriv-insert-here
 
+0.14.0 — 2022-03-25
+===================
+
+Features
+--------
+
+-   `[sc-13426] <https://app.shortcut.com/globus/story/13426>`_
+    Support setting tags when using the ``flow run`` subcommand.
+-   Support batch updates of one or more Runs.
+-   Support updating tags and labels using the ``flow run-update`` subcommand.
+-   Support erasing the list of Run managers and Run monitors using the ``flow run-update`` subcommand.
+    This can be done by specifying an empty string for the value of the ``--run-manager`` and ``--run-monitor`` options.
+
+Bugfixes
+--------
+
+-   `[sc-13664] <https://app.shortcut.com/globus/story/13664/>`_
+    Fix tabular ``run-list`` output.
+-   `[sc-14109] <https://app.shortcut.com/globus/story/14109>`_
+    Mark the ``run-status`` subcommand's ``--flow-id`` option as a mandatory UUID.
+-   `[sc-14127] <https://app.shortcut.com/globus/story/14127>`_
+    Prevent a validation error that occurs when an input schema is not provided to the ``flow deploy`` subcommand.
+
 0.13.1 — 2022-03-02
 ===================